Modify #9 원상복귀 및 수정
- DefaultViewModel 제거
- hashCode 추가 <- ?
